fix(#38): login 페이지의 Client/Server 이중 평가 문제 해결
